Remove docker-sync for console
Docker-sync was introduced to improve performances for OSX developers
working on the console.

It is a pain to use (you need to start a daemon first), and it is not
really reliable (removed files are not detected, though I imagine there
might be an option to fix the problem somehow).

The only really slow thing is the use of dynamic routing ([slug].js),
which is only a small part of the application. For the other pages,
performances are acceptable.

Also, the console doesn't need any access inside the docker network, so
it is possible to run it from the host which doesn't suffer from bad
performances.