README: even better instructions...
Integration requests are all made with the neotaxi account, except a few
cases, even when you're already a "moteur" and you hail a neotaxi taxi.

So neotaxi must be a "moteur" as well.

Maybe neotaxi should just be the complement to the role you don't have,
a "operateur" for moteurs, and vice-versa.