docker-compose: switch to timescaledb
mdillon/postgis have been unmaintained for four years anyway. Production
is currently using Postgres 14, and this environement was stuck to 9.

We use the High Availability image because it's built with Postgis, but
we don't use the HA features.

Use the "trust" authentication method to replicate not typing a password
in the development environment. Obviously not the production settings!